Added VariableScaleDecimalConverter and fixed DecimalConverter

+++ kcbq-connector/src/main/java/com/wepay/kafka/connect/bigquery/convert/logicaltype/KafkaLogicalConverters.java
@@ -70,7 +70,7 @@ public class KafkaLogicalConverters {
     public DecimalConverter() {
       super(Decimal.LOGICAL_NAME,
             Schema.Type.BYTES,
-            LegacySQLTypeName.FLOAT);
+            LegacySQLTypeName.NUMERIC);
     }
 
     @Override
